app/requirements.txt
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
gunicorn
google-genai
python-dotenv
